requirements.txt
streamlit>=1.29
feedparser>=6.0
google-generativeai>=0.7
psycopg2-binary>=2.9
pgvector>=0.2
requests>=2.32
python-dotenv>=1.0
